+++ src/ploomber/dag/DAGConfigurator.py
@@ -7,7 +7,7 @@ unnecessarily complex, custom behavior will be provided via this object.
 This is based in the essence pattern by by Andy Carlson
 http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.81.1898&rep=rep1&type=pdf
 """
-
+from copy import copy
 from ploomber.dag.DAG import DAG
 from ploomber.dag.DAGConfiguration import DAGConfiguration
 
@@ -24,23 +24,26 @@ class DAGConfigurator:
 
     outdated_by_code: whether source code differences make a task outdated
     cache_rendered_status: keep results from dag.render() whenever are needed
-    again (e.g. when calling dag.build()) or compute it again every time
+    again (e.g. when calling dag.build()) or compute it again every time.
+
+    cache_rendered_status: If True, once the DAG is rendered, subsequent calls
+    to render will not do anything (rendering is implicitely called in build,
+    plot, status), otherwise it will always render again.
 
     Examples
     --------
     >>> from ploomber import DAGConfigurator
     >>> configurator = DAGConfigurator()
-    >>> configurator.set_param('outdated_by_code', True)
-    >>> configurator.set_param('cache_rendered_status', False)
+    >>> configurator.cfg.outdated_by_code = True
+    >>> configurator.cfg.cache_rendered_status = False
     >>> dag = configurator.create()
     """
     def __init__(self, cfg=None):
-        self.cfg = cfg or DAGConfiguration()
+        self._cfg = cfg or DAGConfiguration()
 
-    @classmethod
-    def default(cls):
-        cfg = DAGConfiguration.default()
-        return cls(cfg=cfg)
+    @property
+    def cfg(self):
+        return self._cfg
 
     @classmethod
     def from_dict(cls, d):
@@ -51,11 +54,8 @@ class DAGConfigurator:
         """Return a DAG with the given parameters
         """
         dag = DAG()
-        dag._cfg = self.cfg
+        dag._cfg = copy(self.cfg)
 
         # TODO: validate configuration
 
         return dag
-
-    def set_param(self, key, value):
-        return setattr(self.cfg, key, value)